requirements.txt
flask
flask-cors
torch
pillow
numpy
torchvision
requests